pyproject.toml
[tool.poetry]
name = "drf-khynotska"
version = "0.1.0"
description = ""
authors = ["NadiiaKhynotska <nadinyman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
mysqlclient = "^2.2.4"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"





[tool.poetry.group.dev.dependencies]
django-dotenv = "^1.4.2"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
